requirements.txt
gunicorn
Flask
Flask-Mako
tensorflow
keras
numpy
plim
pillow
h5py
